fix compilation for bionic

+++ src/hqc/TimeSeriesView.cc
@@ -349,8 +349,8 @@ std::string TimeSeriesView::changes()
 
   if (mOriginalTimeLimits.t0() != mTimeLimits.t0() or mOriginalTimeLimits.t1() != mTimeLimits.t1()) {
     QDomElement doc_timeshift = doc.createElement(E_TAG_TSHIFT);
-    doc_timeshift.setAttribute(T_ATTR_START, (mTimeLimits.t0() - mOriginalTimeLimits.t0()).hours());
-    doc_timeshift.setAttribute(T_ATTR_END,   (mTimeLimits.t1() - mOriginalTimeLimits.t1()).hours());
+    doc_timeshift.setAttribute(T_ATTR_START, static_cast<int>((mTimeLimits.t0() - mOriginalTimeLimits.t0()).hours()));
+    doc_timeshift.setAttribute(T_ATTR_END,   static_cast<int>((mTimeLimits.t1() - mOriginalTimeLimits.t1()).hours()));
     doc_changes.appendChild(doc_timeshift);
   }
